Autocomplete working + some error handling.
