+++ db_migration_job.mjf
@@ -1,0 +1,266 @@
+[Source]
+Type=304
+TypeName=SQLite
+Server=
+Port=0
+Username=
+Password=
+Database=C:\Users\nico_chemwatch\OneDrive - Ucorp Pty Ltd T A Chemwatch\Documents\Python\Transport of dangerous goods  - Cleanup\heroku-hazmar-parser\instance\hazmat-parser.sqlite
+Schema=
+Driver=
+DriverFileName=
+Engine=
+Charset=
+Collate=
+ConnStr=Provider=MSDASQL;Driver={SQLite Driver};
+ConnMethod=0
+FetchSize=500000
+[Source.Security]
+UseSSL=0
+SSLMode=
+UseAuthentication=0
+SecureAuth=1
+UseEncryptData=0
+Authentication=
+ClientKey=
+ClientCert=
+CACert=
+CAPath=
+Cipher=
+CRL=
+UseDistNames=0
+CN=
+L=
+ST=
+OU=
+O=
+C=
+SysDB=
+DBPassword=
+MDF=
+[Destation]
+Type=18
+TypeName=Oracle
+Server=cwdata.cgzqdcrpzcsp.us-west-1.rds.amazonaws.com
+Port=3306
+Username=6F036E076C03700F600372077C03781F6A036D079503
+Password=65036E073A03460F770366077A037B1F6C0394076303940F62039B078803
+Database=CWDATA
+Schema=NICO_MCGILL
+Driver=
+DriverFileName=
+Engine=
+Charset=
+Collate=
+ConnStr=Provider=OraOLEDB.Oracle;Data Source=%db;User ID=%user;Password=%pwd;Server=%host;Address=%addr;Port=%port;
+ConnMethod=0
+UCF='\ "|=+*&^@!`~:;,.?/<>()[]{}$#%
+UCT=_
+BatchCount=2000
+[Options]
+OverwriteDatabase=0
+[Destation.Security]
+UseSSL=0
+SSLMode=
+UseAuthentication=0
+SecureAuth=1
+UseEncryptData=0
+Authentication=
+ClientKey=
+ClientCert=
+CACert=
+CAPath=
+Cipher=
+CRL=
+UseDistNames=0
+CN=
+L=
+ST=
+OU=
+O=
+C=
+SysDB=
+DBPassword=
+MDF=
+[Tables]
+Count=12
+SName 0=hazmat_table
+SType 0=1024
+DName 0=HAZMAT_TABLE
+TDescr 0=TRIAL
+OnlyStruct 0=-1
+TableIndex 0=-1
+ForeignKey 0=-1
+MakeNameUnique 0=-1
+TO 0=-1
+Name Case 0=-1
+CopyAutoIncrement 0=-1
+Where 0=
+SName 1=label_codes
+SType 1=1024
+DName 1=LABEL_CODES
+TDescr 1=TRIAL
+OnlyStruct 1=-1
+TableIndex 1=-1
+ForeignKey 1=-1
+MakeNameUnique 1=-1
+TO 1=-1
+Name Case 1=-1
+CopyAutoIncrement 1=-1
+Where 1=
+SName 2=other_provisions
+SType 2=1024
+DName 2=OTHER_PROVISIONS
+TDescr 2=TRIAL
+OnlyStruct 2=-1
+TableIndex 2=-1
+ForeignKey 2=-1
+MakeNameUnique 2=-1
+TO 2=-1
+Name Case 2=-1
+CopyAutoIncrement 2=-1
+Where 2=
+SName 3=packaging_exceptions
+SType 3=1024
+DName 3=PACKAGING_EXCEPTIONS
+TDescr 3=TRIAL
+OnlyStruct 3=-1
+TableIndex 3=-1
+ForeignKey 3=-1
+MakeNameUnique 3=-1
+TO 3=-1
+Name Case 3=-1
+CopyAutoIncrement 3=-1
+Where 3=
+SName 4=packaging_instructions
+SType 4=1024
+DName 4=PACKAGING_INSTRUCTIONS
+TDescr 4=TRIAL
+OnlyStruct 4=-1
+TableIndex 4=-1
+ForeignKey 4=-1
+MakeNameUnique 4=-1
+TO 4=-1
+Name Case 4=-1
+CopyAutoIncrement 4=-1
+Where 4=
+SName 5=packaging_requirements
+SType 5=1024
+DName 5=PACKAGING_REQUIREMENTS
+TDescr 5=TRIAL
+OnlyStruct 5=-1
+TableIndex 5=-1
+ForeignKey 5=-1
+MakeNameUnique 5=-1
+TO 5=-1
+Name Case 5=-1
+CopyAutoIncrement 5=-1
+Where 5=
+SName 6=packaging_standards
+SType 6=1024
+DName 6=PACKAGING_STANDARDS
+TDescr 6=TRIAL
+OnlyStruct 6=-1
+TableIndex 6=-1
+ForeignKey 6=-1
+MakeNameUnique 6=-1
+TO 6=-1
+Name Case 6=-1
+CopyAutoIncrement 6=-1
+Where 6=
+SName 7=proper_shipping_names
+SType 7=1024
+DName 7=PROPER_SHIPPING_NAMES
+TDescr 7=TRIAL
+OnlyStruct 7=-1
+TableIndex 7=-1
+ForeignKey 7=-1
+MakeNameUnique 7=-1
+TO 7=-1
+Name Case 7=-1
+CopyAutoIncrement 7=-1
+Where 7=
+SName 8=spec_packaging
+SType 8=1024
+DName 8=SPEC_PACKAGING
+TDescr 8=TRIAL
+OnlyStruct 8=-1
+TableIndex 8=-1
+ForeignKey 8=-1
+MakeNameUnique 8=-1
+TO 8=-1
+Name Case 8=-1
+CopyAutoIncrement 8=-1
+Where 8=
+SName 9=special_provisions
+SType 9=1024
+DName 9=SPECIAL_PROVISIONS
+TDescr 9=TRIAL
+OnlyStruct 9=-1
+TableIndex 9=-1
+ForeignKey 9=-1
+MakeNameUnique 9=-1
+TO 9=-1
+Name Case 9=-1
+CopyAutoIncrement 9=-1
+Where 9=
+SName 10=symbols
+SType 10=1024
+DName 10=SYMBOLS
+TDescr 10=TRIAL
+OnlyStruct 10=-1
+TableIndex 10=-1
+ForeignKey 10=-1
+MakeNameUnique 10=-1
+TO 10=-1
+Name Case 10=-1
+CopyAutoIncrement 10=-1
+Where 10=
+SName 11=tank_cars
+SType 11=1024
+DName 11=TANK_CARS
+TDescr 11=TRIAL
+OnlyStruct 11=-1
+TableIndex 11=-1
+ForeignKey 11=-1
+MakeNameUnique 11=-1
+TO 11=-1
+Name Case 11=-1
+CopyAutoIncrement 11=-1
+Where 11=
+[Field Map 0]
+MapCount=0
+UseFieldMap=0
+[Field Map 1]
+MapCount=0
+UseFieldMap=0
+[Field Map 2]
+MapCount=0
+UseFieldMap=0
+[Field Map 3]
+MapCount=0
+UseFieldMap=0
+[Field Map 4]
+MapCount=0
+UseFieldMap=0
+[Field Map 5]
+MapCount=0
+UseFieldMap=0
+[Field Map 6]
+MapCount=0
+UseFieldMap=0
+[Field Map 7]
+MapCount=0
+UseFieldMap=0
+[Field Map 8]
+MapCount=0
+UseFieldMap=0
+[Field Map 9]
+MapCount=0
+UseFieldMap=0
+[Field Map 10]
+MapCount=0
+UseFieldMap=0
+[Field Map 11]
+MapCount=0
+UseFieldMap=0
